Rename: utils 폴더 통합 및 anims 폴더명 변경
