+++ client/src/animation/AvatarAnims.ts
