+++ client/src/lib/crypto.ts
