+++ client/src/routes/Landing.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom'
-import { getCookie } from '../utils/cookie'
+import { getCookie } from '../lib/cookie'
 import gifImage from '/gif/interverse.gif'
 
 function Landing() {
