+++ client/src/routes/Room.tsx
@@ -4,7 +4,7 @@ import phaserGame from '../PhaserGame'
 import Game from '../scenes/Game'
 import RoomTitle from '../components/RoomTitle'
 import Alert from '../components/Alert/Alert'
-import { getAuthCookie } from '../utils/cookie'
+import { getAuthCookie } from '../lib/cookie'
 import ButtonContainer from '../components/ButtonContainer'
 import VideoModal from '../components/Modals/VideoModal/VideoModal'
 import CreatorModal from '../components/Modals/CreatorModal/CreatorModal'
